feature: Update tab bar at the bottom of the mobile app
