+++ components/Custombar.js
@@ -1,7 +1,7 @@
-import React from 'react';
-import { View, TouchableOpacity, StyleSheet, Dimensions } from 'react-native';
+import React from "react";
+import { View, TouchableOpacity, StyleSheet, Dimensions } from "react-native";
 
-const { width } = Dimensions.get('window');
+const { width } = Dimensions.get("window");
 const baseWidth = 375;
 
 const calcFont = (percent) => {
@@ -18,7 +18,7 @@ const Custombar = ({ state, descriptors, navigation }) => {
 
         const onPress = () => {
           const event = navigation.emit({
-            type: 'tabPress',
+            type: "tabPress",
             target: route.key,
           });
 
@@ -28,16 +28,12 @@ const Custombar = ({ state, descriptors, navigation }) => {
         };
 
         return (
-          <TouchableOpacity
-            key={route.key}
-            onPress={onPress}
-          >
+          <TouchableOpacity key={route.key} onPress={onPress}>
             {options.tabBarIcon &&
               options.tabBarIcon({
                 size: calcFont(34),
-                color: isFocused ? 'green' : 'gray',
-              })
-            }
+                color: "green",
+              })}
           </TouchableOpacity>
         );
       })}
@@ -46,21 +42,18 @@ const Custombar = ({ state, descriptors, navigation }) => {
 };
 
 const styles = StyleSheet.create({
-    bar: {
-        position: 'absolute',   
-        paddingBottom:'7%',     
-        bottom:0,
-        height:'10%',
-        width:'100%',
-        backgroundColor: 'white',    
-        flexDirection: 'row',         
-        alignItems: 'center',         
-        justifyContent: 'space-around',
-        borderTopColor: 'blue',       
-        elevation: 5, 
-
-      },
-     
+  bar: {
+    position: "absolute",
+    paddingBottom: "7%",
+    bottom: 0,
+    height: "10%",
+    width: "100%",
+    backgroundColor: "white",
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-around",
+    borderTopColor: "blue",
+    elevation: 5,
+  },
 });
-  export default Custombar;
-
+export default Custombar;
