+++ components/Tabnav.js
@@ -1,17 +1,15 @@
-import React from 'react';
-import { StyleSheet } from 'react-native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import Home from './tabs/Home.js';
-import About from './tabs/About.js';
-import Learn from './tabs/Learn.js';
-import Curbside from './tabs/Curbside.js';
-import Items from './tabs/Items.js';
-import Icon from 'react-native-vector-icons/MaterialIcons'; 
-import Custombar from './Custombar.js';
-import Fontisto from '@expo/vector-icons/Fontisto';
-import MaterialCommunityIcons from '@expo/vector-icons/MaterialCommunityIcons';
-import MaterialIcons from '@expo/vector-icons/MaterialIcons';
-import Entypo from '@expo/vector-icons/Entypo';
+import React from "react";
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import Home from "./tabs/Home.js";
+import About from "./tabs/About.js";
+import CurbsideDropoff from "./tabs/CurbsideDropoff.js";
+import BarcodeScan from "./tabs/BarcodeScan.js";
+import UserAccount from "./tabs/UserAccount.js";
+import Icon from "react-native-vector-icons/MaterialIcons";
+import Custombar from "./Custombar.js";
+import EvilIcons from "@expo/vector-icons/EvilIcons";
+import MaterialCommunityIcons from "@expo/vector-icons/MaterialCommunityIcons";
+import MaterialIcons from "@expo/vector-icons/MaterialIcons";
 const Tab = createBottomTabNavigator();
 
 function Tabnav() {
@@ -21,49 +19,62 @@ function Tabnav() {
       tabBar={(props) => <Custombar {...props} />}
       initialRouteName="Home"
     >
-      <Tab.Screen 
-        name="About" 
-        component={About} 
+      <Tab.Screen
+        name="Home"
+        component={Home}
         options={{
-          tabBarIcon: ({ size, color }) => <MaterialIcons name="contact-support" size={size} color={color} />,
-          tabBarLabel: 'About',
+          tabBarIcon: ({ size, color }) => (
+            <Icon name="home" size={size} color={color} />
+          ),
+          tabBarLabel: "Home",
         }}
       />
-      <Tab.Screen 
-        name="Learn" 
-        component={Learn} 
+      <Tab.Screen
+        name="Curbside Dropoff"
+        component={CurbsideDropoff}
         options={{
-          tabBarIcon: ({ size, color }) => <Entypo name="open-book" size={size} color={color} />,
-          tabBarLabel: 'Learn',
+          tabBarIcon: ({ size, color }) => (
+            <EvilIcons name="location" size={size} color={color} />
+          ),
+          tabBarLabel: "CurbsideDropoff",
         }}
       />
-      <Tab.Screen 
-        name="Home" 
-        component={Home} 
+      <Tab.Screen
+        name="Barcode Scan"
+        component={BarcodeScan}
         options={{
-          tabBarIcon: ({ size, color }) => <Icon name='home' size={size} color={color} />,
-          tabBarLabel: 'Home',
+          tabBarIcon: ({ size, color }) => (
+            <MaterialCommunityIcons
+              name="barcode-scan"
+              size={size}
+              color={color}
+            />
+          ),
+          tabBarLabel: "BarcodeScan",
         }}
       />
-      <Tab.Screen 
-        name="Curbside" 
-        component={Curbside} 
+      <Tab.Screen
+        name="User Account"
+        component={UserAccount}
         options={{
-          tabBarIcon: ({ size, color }) => <MaterialCommunityIcons name="dump-truck" size={size} color={color} />,
-          tabBarLabel: 'Curbside',
+          tabBarIcon: ({ size, color }) => (
+            <MaterialCommunityIcons name="account" size={size} color={color} />
+          ),
+          tabBarLabel: "UserAccount",
         }}
       />
-      <Tab.Screen 
-        name="Items" 
-        component={Items} 
+      <Tab.Screen
+        name="About"
+        component={About}
         options={{
-          tabBarIcon: ({ size, color }) => <Fontisto name="recycle" size={size} color={color} />,
-          tabBarLabel: 'Items',
+          tabBarIcon: ({ size, color }) => (
+            <MaterialIcons name="contact-support" size={size} color={color} />
+          ),
+          tabBarLabel: "About",
         }}
       />
     </Tab.Navigator>
   );
 }
-
 
 export default Tabnav;
