+++ components/tabs/BarcodeScan.js
@@ -1,0 +1,5 @@
+import { SafeAreaView } from "react-native-safe-area-context";
+
+export default function BarcodeScan() {
+  return <SafeAreaView></SafeAreaView>;
+}
