+++ components/tabs/UserAccount.js
@@ -1,0 +1,5 @@
+import { SafeAreaView } from "react-native-safe-area-context";
+
+export default function UserAccount() {
+  return <SafeAreaView></SafeAreaView>;
+}
